Change photoset to unique component
